fix(expense-claim): resolve bug in description field while saving
### Bug Description
When saving an ERPNext expense claim, the description field in the expenses table was not getting saved. This was due to the "fetch_from" and "fetch_if_empty" properties in the expense_claim_detail.json file.

### Changes Made
Removed the problematic lines:
- "fetch_from": "expense_type.description"
- "fetch_if_empty": 1

### Testing
Tested the fix by creating and saving an expense claim with a description in the expenses table.